Move client join/connect code to new module
